requirements.txt
requests==2.32.5
psutil==7.2.2
pyyaml==6.0.3
certifi==2026.1.4
urllib3==2.6.3
idna==3.11
charset-normalizer==3.4.4
